Move to_tensor_spec with pad value to internal
