+++ tt_metal/api/tt-metalium/experimental/tensor/tensor_apis.hpp
@@ -57,14 +57,6 @@ HostTensor to_row_major_layout(const HostTensor& tensor);
 
 HostTensor to_dtype(const HostTensor& input_tensor, DataType dtype);
 
-// Same convention as HostTensor::from_vector: no default T; pad_value defaults to 0.
-// T is the logical encode / pad element type.
-// Unlike from_vector (T deduced from the buffer), callers must supply T explicitly
-// (to_tensor_spec<float>(t, spec)) or pass a typed pad_value for deduction.
-// Explicit instantiations: float, bfloat16, int32_t, uint32_t, uint16_t, uint8_t (same as from_vector).
-template <typename T>
-HostTensor to_tensor_spec(const HostTensor& tensor, const TensorSpec& dest_spec, T pad_value = 0);
-
 // ======================================================================================
 //                                  Utility functions
 // ======================================================================================
