+++ tt_metal/sources.cmake
@@ -100,6 +100,7 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/experimental/tensor/impl/tensor_impl.hpp
     api/internal/tensor/byte_based_data_transfer_apis.hpp
     api/internal/tensor/host_pad_apis.hpp
+    api/internal/tensor/host_to_tensor_spec_apis.hpp
     api/tt-metalium/experimental/tensor/mesh_tensor.hpp
     api/tt-metalium/experimental/tensor/spec/layout/alignment.hpp
     api/tt-metalium/experimental/tensor/spec/layout/layout.hpp
